+++ app/test/version_test.dart
@@ -8,16 +8,16 @@ void main() {
       expect(AppInfo.version, matches(RegExp(r'^\d+\.\d+\.\d+$')));
     });
 
-    test('AppInfo version should be 0.2.1', () {
-      expect(AppInfo.version, '0.2.1');
+    test('AppInfo version should be 0.2.12', () {
+      expect(AppInfo.version, '0.2.12');
     });
 
     test('AppInfo buildNumber should be valid', () {
       expect(AppInfo.buildNumber, greaterThan(0));
     });
 
-    test('AppInfo buildNumber should be 3', () {
-      expect(AppInfo.buildNumber, 3);
+    test('AppInfo buildNumber should be 1', () {
+      expect(AppInfo.buildNumber, 1);
     });
 
     test('AppInfo userAgent should be valid', () {
@@ -27,7 +27,7 @@ void main() {
     });
 
     test('AppInfo userAgent should match format', () {
-      expect(AppInfo.userAgent, 'V8Ray/0.2.1');
+      expect(AppInfo.userAgent, 'V8Ray/0.2.12');
     });
 
     test('AppInfo should have all required fields', () {